fix: move pinecone integratiosn to pinecone folder
